+++ modules/imgproc/src/histogram.cpp
@@ -1282,8 +1282,8 @@ void cv::calcHist( InputArrayOfArrays images, const std::vector<int>& channels,
 
     CV_OCL_RUN(images.total() == 1 && channels.size() == 1 && images.channels(0) == 1 &&
                channels[0] == 0 && images.isUMatVector() && mask.empty() && !accumulate &&
-               histSize.size() == 1 && histSize[0] == BINS && ranges.size() == 2 &&
-               ranges[0] == 0 && ranges[1] == BINS,
+               histSize.size() == 1 && histSize[0] == (int)BINS && ranges.size() == 2 &&
+               ranges[0] == 0 && ranges[1] == (int)BINS,
                ocl_calcHist(images, hist))
 
     int i, dims = (int)histSize.size(), rsz = (int)ranges.size(), csz = (int)channels.size();
